requirements.txt
Flask
python-dotenv
gunicorn
Flask-Limiter
requests-oauthlib
Flask-JWT-Extended
stripe
pytest
ruff
black